Add openmp support. This is experimental, it seems to work for wrf but not wrfvar
git-svn-id: https://svn-wrf-var.cgd.ucar.edu/trunk@3099 70815b0f-b814-0410-81f7-dfa10a885da6

+++ scripts/da_run_real.ksh
@@ -81,24 +81,25 @@ else
    if [[ -f namelist.output ]]; then
      cp namelist.output $RUN_DIR/namelist.output
    fi
-
-   rm -rf $RUN_DIR/rsl
-   mkdir -p $RUN_DIR/rsl
-   mv rsl* $RUN_DIR/rsl
-   cd $RUN_DIR/rsl
-   for FILE in rsl*; do
-      echo "<HTML><HEAD><TITLE>$FILE</TITLE></HEAD>" > $FILE.html
-      echo "<H1>$FILE</H1><PRE>" >> $FILE.html
-      cat $FILE >> $FILE.html
-      echo "</PRE></BODY></HTML>" >> $FILE.html
-      rm $FILE
-   done
-   cd $RUN_DIR
-
    echo '<A HREF="namelist.output">Namelist output</a>'
-   echo '<A HREF="rsl/rsl.out.0000.html">rsl.out.0000</a>'
-   echo '<A HREF="rsl/rsl.error.0000.html">rsl.error.0000</a>'
-   echo '<A HREF="rsl">Other RSL output</a>'
+
+   if [[ -f rsl.out.0000 ]]; then
+      rm -rf $RUN_DIR/rsl
+      mkdir -p $RUN_DIR/rsl
+      mv rsl* $RUN_DIR/rsl
+      cd $RUN_DIR/rsl
+      for FILE in rsl*; do
+         echo "<HTML><HEAD><TITLE>$FILE</TITLE></HEAD>" > $FILE.html
+         echo "<H1>$FILE</H1><PRE>" >> $FILE.html
+         cat $FILE >> $FILE.html
+         echo "</PRE></BODY></HTML>" >> $FILE.html
+         rm $FILE
+      done
+      echo '<A HREF="rsl/rsl.out.0000.html">rsl.out.0000</a>'
+      echo '<A HREF="rsl/rsl.error.0000.html">rsl.error.0000</a>'
+      echo '<A HREF="rsl">Other RSL output</a>'
+   fi
+   cd $RUN_DIR
 
    echo $(date +'%D %T') "Ended $RC"
 fi
