+++ scripts/da_run_wrf.ksh
@@ -118,8 +118,8 @@ else
       touch wrfnl_go_ahead
    fi
    $RUN_CMD ./wrf.exe
-   grep -q 'SUCCESS COMPLETE WRF' rsl.out.0000 
    RC=$?
+   grep -q 'SUCCESS COMPLETE WRF' rsl.out.0000 
 
    rm -rf $RUN_DIR/rsl
    mkdir -p $RUN_DIR/rsl
@@ -134,11 +134,25 @@ else
    done
    cd $WORK_DIR
    cp namelist.output $RUN_DIR
-
    echo '<A HREF="namelist.output">Namelist output</a>'
-   echo '<A HREF="rsl/rsl.out.0000.html">rsl.out.0000</a>'
-   echo '<A HREF="rsl/rsl.error.0000.html">rsl.error.0000</a>'
-   echo '<A HREF="rsl">Other RSL output</a>'
+
+   if [[ -f rsl.out.0000 ]]; then
+      rm -rf $RUN_DIR/rsl
+      mkdir -p $RUN_DIR/rsl
+      mv rsl* $RUN_DIR/rsl
+      cd $RUN_DIR/rsl
+      for FILE in rsl*; do
+         echo "<HTML><HEAD><TITLE>$FILE</TITLE></HEAD>" > $FILE.html
+         echo "<H1>$FILE</H1><PRE>" >> $FILE.html
+         cat $FILE >> $FILE.html
+         echo "</PRE></BODY></HTML>" >> $FILE.html
+         rm $FILE
+      done
+      echo '<A HREF="rsl/rsl.out.0000.html">rsl.out.0000</a>'
+      echo '<A HREF="rsl/rsl.error.0000.html">rsl.error.0000</a>'
+      echo '<A HREF="rsl">Other RSL output</a>'
+   fi
+
    echo $(date +'%D %T') "Ended $RC"
 fi
